+++ Sources/CodexBar/SettingsStore+Defaults.swift
@@ -760,6 +760,30 @@ extension SettingsStore {
             self.userDefaults.set(newValue.rawValue, forKey: "terminalApp")
         }
     }
+
+    var agentSessionsEnabled: Bool {
+        get { self.defaultsState.agentSessionsEnabled }
+        set {
+            self.defaultsState.agentSessionsEnabled = newValue
+            self.userDefaults.set(newValue, forKey: "agentSessionsEnabled")
+        }
+    }
+
+    var agentSessionsTailscaleEnabled: Bool {
+        get { self.defaultsState.agentSessionsTailscaleEnabled }
+        set {
+            self.defaultsState.agentSessionsTailscaleEnabled = newValue
+            self.userDefaults.set(newValue, forKey: "agentSessionsTailscaleEnabled")
+        }
+    }
+
+    var agentSessionsManualHosts: String {
+        get { self.defaultsState.agentSessionsManualHosts }
+        set {
+            self.defaultsState.agentSessionsManualHosts = newValue
+            self.userDefaults.set(newValue, forKey: "agentSessionsManualHosts")
+        }
+    }
 }
 
 extension SettingsStore {