+++ Sources/CodexBar/SettingsStoreState.swift
@@ -60,4 +60,7 @@ struct SettingsDefaultsState {
     var providersSortedAlphabetically: Bool
     var appLanguageRaw: String?
     var terminalAppRaw: String?
+    var agentSessionsEnabled: Bool
+    var agentSessionsTailscaleEnabled: Bool
+    var agentSessionsManualHosts: String
 }